requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
